Encapsulate system, cpu + ppu in a Nes
This creates a Nes struct to encapsulate a System, Cpu and Ppu
so you aren't responsible for as much when initializing the
state of the emulator.

This should help support more cross-cutting functionality like
tracing logs and should make it more practical to handle more
fiddly synchronization details across all components without
affecting the application.

Conceptually the application instantiates a Nes; loads a
cartridge; inserts the cartridge; resets the Nes; allocates
a framebuffer; repeatedly calls nes.tick() with a framebuffer
that can then be rendered.

This simplifies nes-simple-test a little