simplify oam dma + call ppu step() once per ppu clock
An OAM DMA now happens immediately when writing to $4014 and is assumed
to be unobservable (since the CPU will be stalled).

The plan is to split up the scanline renderer and start emulating on a
per-clock basis instead.

This updates the tick_frame() loop to handle progressing the PPU clock
based on the progress of the CPU clock, as well as handling suspending
the CPU during an OAM DMA.